Correct argument for GetDlgItem
GetDlgItem takes an integer as the identifier of the contol to be
retrieved. buflen is an unsigned integer. Assuming this code works
this just makes the truncation cast explicit. No functionality changes
but buflen *could* be over the limit of an integer. Very unlikey that
there will be a string that is 2^31 long though.

+++ src/main.c
@@ -1095,7 +1095,7 @@ static BOOL CALLBACK InputDlgProc(HWND hDlg, UINT msg, WPARAM wParam, LPARAM lP)
 		if (ctrlId == IDOK) {
 			/* Add user */
 			GetWindowText(
-			    GetDlgItem(hDlg, ID_INPUT_LINE), inBuf->buffer, inBuf->buflen);
+			    GetDlgItem(hDlg, ID_INPUT_LINE), inBuf->buffer, (int)inBuf->buflen);
 			if (strlen(inBuf->buffer) > 0) {
 				EndDialog(hDlg, IDOK);
 			}
